Promote Create button to top level suggestion.
Previously the Create button was not part of the proper suggestions state. This caused several issues:

1. The Create option was only shown if there were suggestions (not what we want in certain circumstances)
2. It wasn’t possible to use keyboard to move to the Create button.
3. a11y concerns regarding having Create option outside of the true suggestions.

We now make the searchSuggestions handler always append a special suggestion to the result set for create. This is a reserved suggestion which is only displayed in the UI when the appropriate conditions are met.

Also fixes bug with un-needed call to `setIsEditingLink( false )` in async `onChange` handler causes invalid test failures.